Refactor ItemEntry to new file
